[#26] Moves gradle related code into own package
Closes #26
